fix: remove incorrect type conversion for status expiration
- Changed profile.setStatusExpiration(statusExpiration.intValue()) to profile.setStatusExpiration(statusExpiration)
- The method expects a Long, not an int
- statusExpiration is already a Long, so no conversion needed

+++ src/main/java/com/trackify/trackify/service/SlackService.java
@@ -84,7 +84,7 @@ public class SlackService {
 
         // Set status expiration if provided
         if (statusExpiration != null) {
-            profile.setStatusExpiration(statusExpiration.intValue());
+            profile.setStatusExpiration(statusExpiration);
         }
 
         UsersProfileSetRequest request = UsersProfileSetRequest.builder()
